+++ bittytax/conv/parsers/__init__.py
@@ -25,6 +25,7 @@ from . import ii
 from . import kucoin
 from . import ledgerlive
 from . import liquid
+from . import nexo
 from . import okex
 from . import poloniex
 from . import qtwallet