+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Crypto.com parser: fix date parser.
 ### Added
 - Conversion tool: added parser for CGTCalculator.
+- Conversion tool: added parser for Nexo.
 ### Changed
 - Hotbit parser: Negative fees are now set to zero.
 - Accounting tool: Drop buy/sell/fee transactions of zero quantity.